requirements.txt
# Apache Airflow
apache-airflow==2.7.0
apache-airflow-providers-amazon==8.7.0

# AWS SDK
boto3==1.29.0
botocore==1.32.0

# Data Processing
pandas==2.1.0
numpy==1.24.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.20

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
